split `utils.ts` into multiple files in `utils/` directory
