+++ package.json
@@ -24,8 +24,8 @@
 			"default": "./dist/logger.js"
 		},
 		"./utils": {
-			"types": "./dist/utils.d.ts",
-			"default": "./dist/utils.js"
+			"types": "./dist/utils/index.d.ts",
+			"default": "./dist/utils/index.js"
 		}
 	},
 	"scripts": {
