+++ src/lib/index.ts
@@ -1,3 +1,2 @@
 export * from './logger.js';
-export * from './url.js';
-export * from './utils.js';
+export * from './utils/index.js';
